refactor: centralize yfinance helpers
